Integrated the functionality of my user tracker plugin [1].
This is useful because sometimes the Twitch IRC servers decide to only send a partial userlist instead of not at all, causing the plugin to ignore the channel as a target for synchronization. This way a more or less accurate list of all the users in the channel can be generated over a while.

[1]: https://github.com/cryzed/User-Tracker